Remove test_string.py
