+++ python/cudf/cudf/tests/series/methods/test_sort_values.py
@@ -1,0 +1,20 @@
+# Copyright (c) 2020-2025, NVIDIA CORPORATION.
+
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing import assert_eq
+
+
+@pytest.mark.parametrize(
+    "data", [["z", "1", "a"], ["c", None, "b"], [None] * 3]
+)
+def test_string_sort(data, ascending):
+    ps = pd.Series(data, dtype="str", name="nice name")
+    gs = cudf.Series(data, dtype="str", name="nice name")
+
+    expect = ps.sort_values(ascending=ascending)
+    got = gs.sort_values(ascending=ascending)
+
+    assert_eq(expect, got)
