--- python/cudf/cudf/tests/test_string.py
@@ -1,355 +1,0 @@
-# Copyright (c) 2018-2025, NVIDIA CORPORATION.
-
-from sys import getsizeof
-
-import numpy as np
-import pandas as pd
-import pytest
-
-import rmm
-
-import cudf
-from cudf.core.buffer import as_buffer
-from cudf.core.column.string import StringColumn
-from cudf.core.index import Index
-from cudf.testing import assert_eq
-from cudf.testing._utils import (
-    assert_exceptions_equal,
-)
-from cudf.utils import dtypes as dtypeutils
-
-
-@pytest.fixture(
-    params=[
-        ["AbC", "de", "FGHI", "j", "kLm"],
-        ["nOPq", None, "RsT", None, "uVw"],
-        [None, None, None, None, None],
-    ],
-    ids=["no_nulls", "some_nulls", "all_nulls"],
-)
-def data(request):
-    return request.param
-
-
-@pytest.fixture(
-    params=[None, [10, 11, 12, 13, 14]], ids=["None_index", "Set_index"]
-)
-def index(request):
-    return request.param
-
-
-@pytest.fixture
-def ps_gs(data, index):
-    ps = pd.Series(data, index=index, dtype="str", name="nice name")
-    gs = cudf.Series(data, index=index, dtype="str", name="nice name")
-    return (ps, gs)
-
-
-@pytest.mark.parametrize("ascending", [True, False])
-def test_string_sort(ps_gs, ascending):
-    ps, gs = ps_gs
-
-    expect = ps.sort_values(ascending=ascending)
-    got = gs.sort_values(ascending=ascending)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "str_data", [[], ["a", "b", "c", "d", "e"], [None, None, None, None, None]]
-)
-@pytest.mark.parametrize("num_keys", [1, 2, 3])
-def test_string_groupby_key(str_data, num_keys):
-    other_data = [1, 2, 3, 4, 5][: len(str_data)]
-
-    pdf = pd.DataFrame()
-    gdf = cudf.DataFrame()
-    for i in range(num_keys):
-        pdf[i] = pd.Series(str_data, dtype="str")
-        gdf[i] = cudf.Series(str_data, dtype="str")
-    pdf["a"] = other_data
-    gdf["a"] = other_data
-
-    expect = pdf.groupby(list(range(num_keys)), as_index=False).count()
-    got = gdf.groupby(list(range(num_keys)), as_index=False).count()
-
-    expect = expect.sort_values([0]).reset_index(drop=True)
-    got = got.sort_values([0]).reset_index(drop=True)
-
-    assert_eq(expect, got, check_dtype=False)
-
-
-@pytest.mark.parametrize(
-    "str_data", [[], ["a", "b", "c", "d", "e"], [None, None, None, None, None]]
-)
-@pytest.mark.parametrize("num_cols", [1, 2, 3])
-@pytest.mark.parametrize("agg", ["count", "max", "min"])
-def test_string_groupby_non_key(str_data, num_cols, agg):
-    other_data = [1, 2, 3, 4, 5][: len(str_data)]
-
-    pdf = pd.DataFrame()
-    gdf = cudf.DataFrame()
-    for i in range(num_cols):
-        pdf[i] = pd.Series(str_data, dtype="str")
-        gdf[i] = cudf.Series(str_data, dtype="str")
-    pdf["a"] = other_data
-    gdf["a"] = other_data
-
-    expect = getattr(pdf.groupby("a", as_index=False), agg)()
-    got = getattr(gdf.groupby("a", as_index=False), agg)()
-
-    expect = expect.sort_values(["a"]).reset_index(drop=True)
-    got = got.sort_values(["a"]).reset_index(drop=True)
-
-    if agg in ["min", "max"] and len(expect) == 0 and len(got) == 0:
-        for i in range(num_cols):
-            expect[i] = expect[i].astype("str")
-
-    assert_eq(expect, got, check_dtype=False)
-
-
-def test_string_groupby_key_index():
-    str_data = ["a", "b", "c", "d", "e"]
-    other_data = [1, 2, 3, 4, 5]
-
-    pdf = pd.DataFrame()
-    gdf = cudf.DataFrame()
-    pdf["a"] = pd.Series(str_data, dtype="str")
-    gdf["a"] = cudf.Series(str_data, dtype="str")
-    pdf["b"] = other_data
-    gdf["b"] = other_data
-
-    expect = pdf.groupby("a", sort=True).count()
-    got = gdf.groupby("a", sort=True).count()
-
-    assert_eq(expect, got, check_dtype=False)
-
-
-@pytest.mark.parametrize("scalar", ["a", None])
-def test_string_set_scalar(scalar):
-    pdf = pd.DataFrame()
-    pdf["a"] = [1, 2, 3, 4, 5]
-    gdf = cudf.DataFrame.from_pandas(pdf)
-
-    pdf["b"] = "a"
-    gdf["b"] = "a"
-
-    assert_eq(pdf["b"], gdf["b"])
-    assert_eq(pdf, gdf)
-
-
-def test_string_index():
-    rng = np.random.default_rng(seed=0)
-    pdf = pd.DataFrame(rng.random(size=(5, 5)))
-    gdf = cudf.DataFrame.from_pandas(pdf)
-    stringIndex = ["a", "b", "c", "d", "e"]
-    pdf.index = stringIndex
-    gdf.index = stringIndex
-    assert_eq(pdf, gdf)
-    stringIndex = np.array(["a", "b", "c", "d", "e"])
-    pdf.index = stringIndex
-    gdf.index = stringIndex
-    assert_eq(pdf, gdf)
-    stringIndex = Index(["a", "b", "c", "d", "e"], name="name")
-    pdf.index = stringIndex.to_pandas()
-    gdf.index = stringIndex
-    assert_eq(pdf, gdf)
-    stringIndex = cudf.Index._from_column(
-        cudf.core.column.as_column(["a", "b", "c", "d", "e"]), name="name"
-    )
-    pdf.index = stringIndex.to_pandas()
-    gdf.index = stringIndex
-    assert_eq(pdf, gdf)
-
-
-@pytest.mark.parametrize(
-    "item",
-    [
-        ["Cbe", "cbe", "CbeD", "Cb", "ghi", "Cb"],
-        ["a", "a", "a", "a", "A"],
-        ["A"],
-        ["abc", "xyz", None, "ab", "123"],
-        [None, None, "abc", None, "abc"],
-    ],
-)
-def test_string_unique(item):
-    ps = pd.Series(item)
-    gs = cudf.Series(item)
-    # Pandas `unique` returns a numpy array
-    pres = pd.Series(ps.unique())
-    # cudf returns a cudf.Series
-    gres = gs.unique()
-    assert_eq(pres, gres)
-
-
-def test_string_equality():
-    data1 = ["b", "c", "d", "a", "c"]
-    data2 = ["a", None, "c", "a", "c"]
-
-    ps1 = pd.Series(data1)
-    ps2 = pd.Series(data2)
-    gs1 = cudf.Series(data1)
-    gs2 = cudf.Series(data2)
-
-    expect = ps1 == ps2
-    got = gs1 == gs2
-
-    assert_eq(expect, got.fillna(False))
-
-    expect = ps1 == "m"
-    got = gs1 == "m"
-
-    assert_eq(expect, got.fillna(False))
-
-    ps1 = pd.Series(["a"])
-    gs1 = cudf.Series(["a"])
-
-    expect = ps1 == "m"
-    got = gs1 == "m"
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "lhs",
-    [
-        ["Cbe", "cbe", "CbeD", "Cb", "ghi", "Cb"],
-        ["abc", "xyz", "a", "ab", "123", "097"],
-    ],
-)
-@pytest.mark.parametrize(
-    "rhs",
-    [
-        ["Cbe", "cbe", "CbeD", "Cb", "ghi", "Cb"],
-        ["a", "a", "a", "a", "A", "z"],
-    ],
-)
-def test_string_binary_op_add(lhs, rhs):
-    pds = pd.Series(lhs) + pd.Series(rhs)
-    gds = cudf.Series(lhs) + cudf.Series(rhs)
-
-    assert_eq(pds, gds)
-
-
-def test_string_no_children_properties():
-    empty_col = StringColumn(
-        as_buffer(rmm.DeviceBuffer(size=0)),
-        size=0,
-        dtype=np.dtype("object"),
-        children=(),
-    )
-    assert empty_col.base_children == ()
-    assert empty_col.base_size == 0
-
-    assert empty_col.children == ()
-    assert empty_col.size == 0
-
-    assert getsizeof(empty_col) >= 0  # Accounts for Python GC overhead
-
-
-def test_string_table_view_creation():
-    data = ["hi"] * 25 + [None] * 2027
-    psr = pd.Series(data)
-    gsr = cudf.Series.from_pandas(psr)
-
-    expect = psr[:1]
-    got = gsr[:1]
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data,dtype",
-    [
-        (["0.1", "10.2", "10.876"], "float"),
-        (["-0.1", "10.2", "+10.876"], "float"),
-        (["1", "10.2", "10.876"], "float32"),
-        (["+123", "6344556789", "0"], "int"),
-        (["+123", "6344556789", "0"], "uint64"),
-        (["+123", "6344556789", "0"], "float"),
-        (["0.1", "-10.2", "10.876", None], "float"),
-    ],
-)
-@pytest.mark.parametrize("obj_type", [None, "str", "category"])
-def test_string_typecast(data, obj_type, dtype):
-    psr = pd.Series(data, dtype=obj_type)
-    gsr = cudf.Series(data, dtype=obj_type)
-
-    expect = psr.astype(dtype=dtype)
-    actual = gsr.astype(dtype=dtype)
-    assert_eq(expect, actual)
-
-
-@pytest.mark.parametrize(
-    "data,dtype",
-    [
-        (["0.1", "10.2", "10.876"], "int"),
-        (["1", "10.2", "+10.876"], "int"),
-        (["abc", "1", "2", " "], "int"),
-        (["0.1", "10.2", "10.876"], "uint64"),
-        (["1", "10.2", "+10.876"], "uint64"),
-        (["abc", "1", "2", " "], "uint64"),
-        ([" ", "0.1", "2"], "float"),
-        ([""], "int"),
-        ([""], "uint64"),
-        ([" "], "float"),
-        (["\n"], "int"),
-        (["\n"], "uint64"),
-        (["0.1", "-10.2", "10.876", None], "int"),
-        (["0.1", "-10.2", "10.876", None], "uint64"),
-        (["0.1", "-10.2", "10.876", None, "ab"], "float"),
-        (["+", "-"], "float"),
-        (["+", "-"], "int"),
-        (["+", "-"], "uint64"),
-        (["1++++", "--2"], "float"),
-        (["1++++", "--2"], "int"),
-        (["1++++", "--2"], "uint64"),
-        (["++++1", "--2"], "float"),
-        (["++++1", "--2"], "int"),
-        (["++++1", "--2"], "uint64"),
-    ],
-)
-@pytest.mark.parametrize("obj_type", [None, "str", "category"])
-def test_string_typecast_error(data, obj_type, dtype):
-    psr = pd.Series(data, dtype=obj_type)
-    gsr = cudf.Series(data, dtype=obj_type)
-
-    assert_exceptions_equal(
-        lfunc=psr.astype,
-        rfunc=gsr.astype,
-        lfunc_args_and_kwargs=([dtype],),
-        rfunc_args_and_kwargs=([dtype],),
-    )
-
-
-def test_string_int_to_ipv4():
-    gsr = cudf.Series([0, None, 0, 698875905, 2130706433, 700776449]).astype(
-        "uint32"
-    )
-    expected = cudf.Series(
-        ["0.0.0.0", None, "0.0.0.0", "41.168.0.1", "127.0.0.1", "41.197.0.1"]
-    )
-
-    got = cudf.Series._from_column(gsr._column.int2ip())
-
-    assert_eq(expected, got)
-
-
-@pytest.mark.parametrize(
-    "dtype", sorted(list(dtypeutils.NUMERIC_TYPES - {"uint32"}))
-)
-def test_string_int_to_ipv4_dtype_fail(dtype):
-    gsr = cudf.Series([1, 2, 3, 4, 5]).astype(dtype)
-    with pytest.raises(TypeError):
-        gsr._column.int2ip()
-
-
-def test_string_slice_with_mask():
-    actual = cudf.Series(["hi", "hello", None])
-    expected = actual[0:3]
-
-    assert actual._column.base_size == 3
-    assert_eq(actual._column.base_size, expected._column.base_size)
-    assert_eq(actual._column.null_count, expected._column.null_count)
-
-    assert_eq(actual, expected)
